Migrate release pipeline from azure to github actions
